Use ExecResult return type to avoid mutating context.hasMissingField.
If we want to cache the results of the execute{SelectionSet,Field,...}
functions, then context.hasMissingField cannot be mutated as a side effect
of readStoreResolver, because readStoreResolver might not be called again,
thanks to the caching.

Instead, we need to include any information we want to be cached in the
return values of the functions that participate in the caching.

Since the hasMissingField property is ultimately negated to become the
.complete property of the object returned by diffQueryAgainstStore, I
decided to use the .complete property everywhere.